Extend expiration date when keeping same level, calculate from now when changing levels. #721

+++ includes/gateways/class-rcp-payment-gateway-manual.php
@@ -33,9 +33,13 @@ class RCP_Payment_Gateway_Manual extends RCP_Payment_Gateway {
 
 		$member = new RCP_Member( $this->user_id );
 
-		delete_user_meta( $member->ID, 'rcp_pending_expiration_date' );
-
-		$expiration = $member->calculate_expiration( true );
+		$old_level = get_user_meta( $member->ID, '_rcp_old_subscription_id', true );
+		if ( ! empty( $old_level ) && $old_level == $this->subscription_id ) {
+			$expiration = $member->calculate_expiration();
+		} else {
+			delete_user_meta( $member->ID, 'rcp_pending_expiration_date' );
+			$expiration = $member->calculate_expiration( true );
+		}
 
 		$member->renew( false, 'pending', $expiration );
